requirements.txt
dj-database-url==0.5.0
Django==1.11
django-filter==2.2.0
django-mysql==3.3.0
djangorestframework==3.7.0
gunicorn==20.0.4
mysqlclient==1.4.6
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyMySQL==0.9.3
pytz==2019.3
whitenoise==3.3.1
